graph database

+++ fuellogger_project/src/no/appfortress/graph/MockData.java
@@ -5,7 +5,7 @@ import java.util.Random;
 public class MockData {
 
 	// x is the day number, 0, 1, 2, 3
-	public static Point getDataFromReceiver(int x, int y)
+	public static Point getDataFromReceiver(double x, double y)
 	{
 		return new Point(x, y);
 	}
